+++ Dtos/OrderItemDto.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }              // Có thể cần khi update
         public int OrderId { get; set; }         // Liên kết đến Order
         public int ProductId { get; set; }       // Sản phẩm trong đơn
-        public int Quantity { get; set; }        // Số lượng
+        public decimal Quantity { get; set; }        // Số lượng
         public decimal Price { get; set; }       // Giá từng sản phẩm
         public int? EnterpriseId { get; set; }  // EnterpriseId của sản phẩm (để frontend filter)
         public string? EnterpriseName { get; set; } // Tên enterprise (để frontend hiển thị)
